+++ rust/src/ftp/ftp.rs
@@ -19,6 +19,7 @@ use std;
 use std::ffi::CString;
 use std::os::raw::{c_char, c_int, c_void};
 
+use crate::core::*;
 use crate::conf::{conf_get, get_memval};
 use crate::ftp::constant::*;
 use lazy_static::lazy_static;
@@ -95,14 +96,6 @@ lazy_static! {
         FtpCommand::new("USER", FtpRequestCommand::FTP_COMMAND_USER),
         FtpCommand::new("UNKNOWN", FtpRequestCommand::FTP_COMMAND_UNKNOWN),
     ];
-}
-
-/// cbindgen:ignore
-extern "C" {
-    pub fn MpmAddPatternCI(
-        ctx: *const c_void, pat: *const libc::c_char, pat_len: c_int, _offset: c_int,
-        _depth: c_int, id: c_int, rule_id: c_int, _flags: c_int,
-    ) -> c_void;
 }
 
 #[allow(non_snake_case)]
